dsBase: 'DataSHIELD' Server Side Base Functions

Base 'DataSHIELD' functions for the server side. 'DataSHIELD' is a software package which allows you to do non-disclosive federated analysis on sensitive data. 'DataSHIELD' analytic functions have been designed to only share non disclosive summary statistics, with built in automated output checking based on statistical disclosure control. With data sites setting the threshold values for the automated output checks. For more details, see 'citation("dsBase")'.

Author

Maintainer: Stuart Wheater stuart.wheater@arjuna.com (ORCID)

Authors:

Stuart Wheater stuart.wheater@arjuna.com (ORCID)

Paul Burton (ORCID)

Rebecca Wilson (ORCID)

Olly Butters (ORCID)

Patricia Ryser-Welch (ORCID)

Alex Westerberg

Leire Abarrategui

Roberto Villegas-Diaz (ORCID)

Demetris Avraam (ORCID)

Yannick Marcon yannick.marcon@obiba.org (ORCID)

Tom Bishop

Amadou Gaye (ORCID)

Xavier Escribà-Montagut (ORCID)