tableDS is the first of two serverside aggregate functions called by ds.table

creates 1-dimensional, 2-dimensional and 3-dimensional tables using the table function in native R.

Usage

tableDS(
  rvar.transmit,
  cvar.transmit,
  stvar.transmit,
  rvar.all.unique.levels.transmit,
  cvar.all.unique.levels.transmit,
  stvar.all.unique.levels.transmit,
  exclude.transmit,
  useNA.transmit,
  force.nfilter.transmit
)

Arguments

rvar.transmit

is a character string (in inverted commas) specifying the name of the variable defining the rows in all of the 2 dimensional tables that form the output. Fully specified by <rvar> argument in ds.table. For more information see help for ds.table

cvar.transmit

is a character string specifying the name of the variable defining the columns in all of the 2 dimensional tables that form the output. Fully specified by <cvar> argument in ds.table. For more information see help for ds.table

stvar.transmit

is a character string specifying the name of the variable that indexes the separate two dimensional tables in the output if the call specifies a 3 dimensional table. Fully specified by <stvar> argument in ds.table. For more information see help for ds.table

rvar.all.unique.levels.transmit

is a character string containing all unique level in rvar, across the studies, separated by ','.

cvar.all.unique.levels.transmit

is a character string containing all unique level in cvar, across the studies, separated by ','.

stvar.all.unique.levels.transmit

is a character string containing all unique level in stvar, across the studies, separated by ','.

exclude.transmit

for information see help on <exclude> argument of ds.table. Fully specified by <exclude> argument of ds.table

useNA.transmit

for information see help on <useNA> argument of ds.table. Fully specified by <useNA> argument of ds.table

force.nfilter.transmit

for information see help on <force.nfilter> argument of ds.table. Fully specified by <force.nfilter> argument of ds.table

Value

For information see help for ds.table

Details

this serverside function is the workhorse of ds.table - creating the table requested in the format specified by ds.table. For more information see help for ds.table in DataSHIELD and the table function in native R.

Author

Paul Burton for DataSHIELD Development Team, 13/11/2019